+++ common/scripts/definitions/maps/debug.ts
@@ -112,9 +112,9 @@ export const SingleBuildMap:MapDef={
         },
     },
     gen_callback(map) {
-        const def=map.game.definitions.buildings.getFromString("shed")
+        //const def=map.game.definitions.buildings.getFromString("shed")
         //const def=map.game.definitions.buildings.getFromString("bunker_1")
-        //const def=map.game.definitions.buildings.getFromString("small_house_1")
+        const def=map.game.definitions.buildings.getFromString("small_house_1")
         //const def=map.game.definitions.buildings.getFromString(`${random.choose(["yellow","blue","red","green"])}_container_${random.int(1,2)}`)
 
         const b=map.game.map.add_building(def)